Make agent_data and base_agent_data generic classes
